Code refactor for better maintaince
